requirements.txt
requests
websocket_client
./dist/NorenRestApiPy-0.0.22-py2.py3-none-any.whl
pandas
pyyaml
python-dotenv
pynput